+++ project/job/admin/src/main/java/com/xxl/job/admin/dao/XxlJobUserDao.java
@@ -28,6 +28,6 @@ public interface XxlJobUserDao {
 
     public int update(XxlJobUser xxlJobUser);
 
-    public int delete(@Param("id") int id);
+    public int delete(@Param("id") Long id);
 
 }
